pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ascend"
version = "0.1.0"
description = "AI-powered engineering management CLI"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "textual>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
ascend = "ascend.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ascend"]

[tool.pytest.ini_options]
testpaths = ["tests"]
